tests: moved tests into __tests__ folder
