+++ frontend/app/src/pages/superadmin/accessDenied/__tests__/AccessDenied.Spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import AdminAccessDenied from './index.tsx';
+import AdminAccessDenied from '../index.tsx';
 
 describe('AdminAccessDenied', () => {
   test("renders Access Denied message, You don't have access message, and Go Back button", () => {
