+++ frontend/app/src/pages/superadmin/header/__tests__/SuperAdminHeader.spec.tsx
@@ -3,10 +3,10 @@ import { render, screen, within, act } from '@testing-library/react';
 import moment from 'moment';
 import nock from 'nock';
 import React from 'react';
-import { setupStore } from '../../../__test__/__mockData__/setupStore';
-import { user } from '../../../__test__/__mockData__/user';
-import { mockUsehistory } from '../../../__test__/__mockFn__/useHistory';
-import { Header } from './';
+import { setupStore } from '../../../../__test__/__mockData__/setupStore';
+import { user } from '../../../../__test__/__mockData__/user';
+import { mockUsehistory } from '../../../../__test__/__mockFn__/useHistory';
+import { Header } from '../';
 
 beforeAll(() => {
   nock.disableNetConnect();
