+++ frontend/app/src/pages/superadmin/statistics/__tests__/Statistics.spec.tsx
@@ -2,10 +2,10 @@ import '@testing-library/jest-dom';
 import { render, screen, waitFor } from '@testing-library/react';
 import nock from 'nock';
 import React from 'react';
-import { setupStore } from '../../../__test__/__mockData__/setupStore';
-import { user } from '../../../__test__/__mockData__/user';
-import { mockUsehistory } from '../../../__test__/__mockFn__/useHistory';
-import { Statistics } from './';
+import { setupStore } from '../../../../__test__/__mockData__/setupStore';
+import { user } from '../../../../__test__/__mockData__/user';
+import { mockUsehistory } from '../../../../__test__/__mockFn__/useHistory';
+import { Statistics } from '../';
 
 beforeAll(() => {
   nock.disableNetConnect();
